+++ packages/sit-onyx/src/components/OnyxCheckbox/OnyxCheckbox.ct.tsx
@@ -94,29 +94,44 @@ test("should render indeterminate", async ({ mount, makeAxeBuilder }) => {
   await expect(component).toHaveScreenshot("indeterminate.png");
 });
 
-test("should render disabled", async ({ mount, makeAxeBuilder }) => {
-  const testCases = ["unchecked", "checked", "indeterminate"] as const;
+test("should render required", async ({ mount, makeAxeBuilder }) => {
+  // ARRANGE
+  const component = await mount(<OnyxCheckbox label="Required" required />);
 
-  for (const testCase of testCases) {
+  // ACT
+  const accessibilityScanResults = await makeAxeBuilder().analyze();
+
+  // ASSERT
+  expect(accessibilityScanResults.violations).toEqual([]);
+  await expect(component).toHaveScreenshot("required.png");
+});
+
+const disabledTestCases = [
+  { name: "unchecked" },
+  { name: "checked", modelValue: true },
+  { name: "indeterminate", indeterminate: true },
+];
+for (const testCase of disabledTestCases) {
+  test(`should render disabled (${testCase.name})`, async ({ mount, makeAxeBuilder }) => {
     // ARRANGE
     const component = await mount(
       <div style="display: grid; width: max-content;">
         <OnyxCheckbox
-          label={`Disabled ${testCase}`}
-          modelValue={testCase === "checked"}
-          indeterminate={testCase === "indeterminate"}
+          label={`Disabled ${testCase.name}`}
+          modelValue={testCase.modelValue}
+          indeterminate={testCase.indeterminate}
           disabled
         />
         <OnyxCheckbox
           label="Hover"
-          modelValue={testCase === "checked"}
-          indeterminate={testCase === "indeterminate"}
+          modelValue={testCase.modelValue}
+          indeterminate={testCase.indeterminate}
           disabled
         />
         <OnyxCheckbox
           label="Focus visible"
-          modelValue={testCase === "checked"}
-          indeterminate={testCase === "indeterminate"}
+          modelValue={testCase.modelValue}
+          indeterminate={testCase.indeterminate}
           disabled
         />
       </div>,
@@ -140,6 +155,61 @@ test("should render disabled", async ({ mount, makeAxeBuilder }) => {
     }
 
     // ASSERT
-    await expect(component).toHaveScreenshot(`disabled-${testCase}.png`);
-  }
-});
+    await expect(component).toHaveScreenshot(`disabled-${testCase.name}.png`);
+  });
+}
+
+const invalidTestCases = [
+  { name: "unchecked" },
+  { name: "unchecked (disabled)", disabled: true },
+  { name: "indeterminate", indeterminate: true },
+  { name: "indeterminate (disabled)", indeterminate: true, disabled: true },
+];
+for (const testCase of invalidTestCases) {
+  test(`should render invalid (${testCase.name})`, async ({ mount, makeAxeBuilder }) => {
+    // ARRANGE
+    const component = await mount(
+      <div style="display: grid; width: max-content;">
+        <OnyxCheckbox
+          label={`Invalid ${testCase.name}`}
+          indeterminate={testCase.indeterminate}
+          disabled={testCase.disabled}
+          required
+        />
+        <OnyxCheckbox
+          label="Hover"
+          indeterminate={testCase.indeterminate}
+          disabled={testCase.disabled}
+          required
+        />
+        <OnyxCheckbox
+          label="Focus visible"
+          indeterminate={testCase.indeterminate}
+          disabled={testCase.disabled}
+          required
+        />
+      </div>,
+    );
+
+    // ACT
+    const accessibilityScanResults = await makeAxeBuilder().analyze();
+
+    // ASSERT
+    expect(accessibilityScanResults.violations).toEqual([]);
+
+    const checkboxes = await component.getByRole("checkbox").all();
+
+    // invalid only shows if checkbox is touched
+    for (const checkbox of checkboxes) {
+      await checkbox.focus();
+      await checkbox.blur();
+    }
+
+    // ACT
+    await checkboxes[1].hover();
+    await checkboxes[2].focus();
+
+    // ASSERT
+    await expect(component).toHaveScreenshot(`invalid-${testCase.name}.png`);
+  });
+}