+++ packages/sit-onyx/src/components/OnyxRadioButtonGroup/OnyxRadioButtonGroup.ct.tsx
@@ -6,9 +6,9 @@ const EXAMPLE_OPTIONS: SelectionOption<string>[] = [
   { label: "dummy.1", value: "1", id: "1" },
   { label: "dummy.2", value: "2", id: "2" },
   { label: "dummy.3", value: "3", id: "3" },
-  { label: "Loading", value: "4", id: "4", isLoading: true },
-  { label: "Readonly", value: "5", id: "5", isReadonly: true },
-  { label: "Disabled", value: "6", id: "6", isDisabled: true },
+  { label: "Loading", value: "4", id: "4", loading: true },
+  { label: "Readonly", value: "5", id: "5", readonly: true },
+  { label: "Disabled", value: "6", id: "6", disabled: true },
 ];
 
 test("should display correctly", async ({ mount, makeAxeBuilder, page }) => {
@@ -16,7 +16,7 @@ test("should display correctly", async ({ mount, makeAxeBuilder, page }) => {
   await mount(
     <OnyxRadioButtonGroup
       options={EXAMPLE_OPTIONS}
-      label="radio group label"
+      headline="radio group label"
       name="radio-selection"
     />,
   );
@@ -40,7 +40,7 @@ test("should display correctly when preselected", async ({ mount, makeAxeBuilder
   await mount(
     <OnyxRadioButtonGroup
       options={EXAMPLE_OPTIONS}
-      label="radio group label"
+      headline="radio group label"
       name="radio-selection"
       modelValue={EXAMPLE_OPTIONS[0]}
       onUpdate:modelValue={(u) => updates.push(u)}