+++ apps/alpha-test-app/CHANGELOG.md
@@ -1,5 +1,33 @@
 # alpha-test-app
 
+## 0.0.1-alpha.12
+
+### Patch Changes
+
+- Updated dependencies [1de4414]
+  - sit-onyx@1.0.0-alpha.15
+
+## 0.0.1-alpha.11
+
+### Patch Changes
+
+- Updated dependencies [c605dbb]
+  - sit-onyx@1.0.0-alpha.14
+
+## 0.0.1-alpha.10
+
+### Patch Changes
+
+- Updated dependencies [1c52657]
+  - sit-onyx@1.0.0-alpha.13
+
+## 0.0.1-alpha.9
+
+### Patch Changes
+
+- Updated dependencies [95b6e75]
+  - sit-onyx@1.0.0-alpha.12
+
 ## 0.0.1-alpha.8
 
 ### Patch Changes